Version 4.7b where .int files can be either .rom or .bin and Nintellivision will auto-detect. This allows you to rename all Intellivision game ROMs to .int so that Twilight Menu will recognize them. Also slight tweak to Intellivoice games to speed them up 1%

+++ arm9/source/emucore/SP0256.h
@@ -67,7 +67,7 @@ extern  INT32 currentBits;
 //registers
 extern  UINT16 pc;
 extern  UINT16 stack;
-extern  INT32 mode;
+extern  UINT8  mode;
 extern  INT32 repeatPrefix;
 extern  INT32 command;
